Moved Tests for shuffle to another test class 'ChangeShuffleStateTest'
